Deduplicate legacy const generics logic
