+++ compiler/rustc_ast_lowering/src/expr.rs
@@ -27,7 +27,7 @@ use crate::diagnostics::{
 };
 use crate::{
     AllowReturnTypeNotation, GenericArgsMode, ImplTraitContext, ImplTraitPosition, LoweringContext,
-    ParamMode, ResolverAstLoweringExt, TryBlockScope,
+    ParamMode, TryBlockScope,
 };
 
 pub(super) struct WillCreateDefIdsVisitor;
@@ -211,8 +211,7 @@ impl<'hir> LoweringContext<'_, 'hir> {
                 }
                 ExprKind::Tup(elts) => hir::ExprKind::Tup(self.lower_exprs(elts)),
                 ExprKind::Call(f, args) => {
-                    if let Some(legacy_args) = self.resolver.legacy_const_generic_args(f, self.tcx)
-                    {
+                    if let Some(legacy_args) = self.owner.legacy_const_generic_args(f, self.tcx) {
                         self.lower_legacy_const_generics((**f).clone(), args.clone(), &legacy_args)
                     } else {
                         let f = self.lower_expr(f);
